+++ src/main/java/com/prgrms/mukvengers/global/security/token/dto/jwt/JwtAuthenticationToken.java
@@ -1,4 +1,4 @@
-package com.prgrms.mukvengers.global.security.jwt;
+package com.prgrms.mukvengers.global.security.token.dto.jwt;
 
 import java.util.Collection;
